refactor: styled components structre for the about ticker details and prev close components
