+++ apps/nasdaq/src/components/about/IndustryInfo.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+const Info = styled.p`
+  color: var(--color-main);
+  font-weight: 700;
+  line-height: 1.5rem;
+`;
+
+export default Info;
