+++ apps/nasdaq/src/components/about/IndustryTitle.ts
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const Title = styled.p`
+  color: var(--color-main);
+  font-size: 1.125rem;
+  margin-block-end: 0.25rem;
+`;
+export default Title;
