+++ apps/nasdaq/src/components/about/IndustryWrapper.ts
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+const IndustryWrapper = styled.div`
+  margin-block-end: 1rem;
+`;
+export default IndustryWrapper;
