+++ apps/nasdaq/src/components/about/ViewWebsiteLink.ts
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+const ViewWebsite = styled.a`
+  color: var(--color-main);
+  font-weight: 700;
+  text-decoration: underline;
+  position: absolute;
+  right: 2rem;
+  top: 1rem;
+`;
+
+export default ViewWebsite;
