+++ apps/nasdaq/src/styled-components/BackButton.ts
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+const Button = styled.button`
+  background-color: transparent;
+  border: none;
+  cursor: pointer;
+  padding: 0;
+`;
+
+export default Button;
