+++ apps/nasdaq/src/styled-components/BackToHomeButtonWrapper.ts
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const BackToHomeButtonWrapper = styled.div`
+  position: absolute;
+  left: 2rem;
+  top: 2.5rem;
+`;
+export default BackToHomeButtonWrapper;
