+++ apps/nasdaq/src/styled-components/Image.ts
