+++ apps/nasdaq/src/styled-components/Search.ts
