+++ apps/nasdaq/src/styled-components/StockArticleWrapper.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+const StockAbout = styled.article`
+  padding-block: 1rem;
+  padding-inline: var(--padding-start);
+  position: relative;
+`;
+
+export default StockAbout;
