+++ apps/nasdaq/src/styled-components/StockDetailsHeadingWrapper.ts
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const StockDetailsHeadingWrapper = styled.span`
+  display: block;
+  padding-block: 1rem;
+  padding-inline: var(--padding-start);
+`;
+export default StockDetailsHeadingWrapper;
