+++ apps/nasdaq/src/styled-components/StockDetailsSections.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+const StockDetailsSections = styled.section`
+  background-color: var(--color-secondary-background);
+  height: 100vh;
+  padding-block-start: 2rem;
+`;
+
+export default StockDetailsSections;
