+++ apps/nasdaq/src/styled-components/StockPrevCloseWrapper.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+const StockPrevCloseWrapper = styled.article`
+  background-color: var(--color-background);
+  padding-block: 1rem;
+  padding-inline: var(--padding-start);
+`;
+
+export default StockPrevCloseWrapper;
