+++ apps/nasdaq/src/styled-components/TickerLogo.ts
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+const TickerLogo = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: var(--color-background);
+  border-radius: 8px;
+  width: 60px;
+  height: 60px;
+  margin: auto;
+`;
+
+export default TickerLogo;
