+++ apps/nasdaq/src/styled-components/TickerName.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+const TickerName = styled.h1`
+  color: var(--color-main);
+`;
+
+export default TickerName;
