+++ apps/nasdaq/src/styled-components/about-ticker/Heading.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+const H2 = styled.h2`
+  color: var(--color-main);
+  margin-block-end: 1rem;
+`;
+export default H2;
