+++ apps/nasdaq/src/styled-components/prev-close-card/Heading.ts
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const H2 = styled.h2`
+  color: var(--color-main);
+  margin-block-end: 1rem;
+`;
+
+export default H2;
