+++ apps/nasdaq/src/styled-components/prev-close-card/StatisticValue.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+const StatisticsValue = styled.span`
+  color: var(--color-main);
+  font-size: 2rem;
+  font-weight: 700;
+`;
+
+export default StatisticsValue;
