+++ apps/nasdaq/src/styled-components/prev-close-card/StatisticsTitle.ts
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const StatisticsTitle = styled.p`
+  color: var(--color-main);
+  font-size: 1.25rem;
+  width: 100%;
+`;
+export default StatisticsTitle;
