+++ apps/nasdaq/src/styled-components/prev-close-card/StatisticsWrapper.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+const StatisticsWrapper = styled.div`
+  display: flex;
+  flex-wrap: wrap;
+`;
+export default StatisticsWrapper;
